Feat: Service, Repository 분리

+++ src/main/java/com/ll/exam/article/dto/ArticleDto.java
@@ -1,21 +1,21 @@
 package com.ll.exam.article.dto;
 
 public class ArticleDto {
-    private int id;
+    private long id;
     private String title;
     private String body;
 
-    public ArticleDto(int id, String title, String body) {
+    public ArticleDto(long id, String title, String body) {
         this.id = id;
         this.title = title;
         this.body = body;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
